+++ python-engine-test/src/test/java/io/deephaven/engine/table/impl/select/TestConditionFilter.java
@@ -16,7 +16,7 @@ import io.deephaven.compilertools.CompilerTools;
 import io.deephaven.engine.table.Table;
 import io.deephaven.engine.table.lang.QueryScopeParam;
 import io.deephaven.engine.table.lang.QueryScope;
-import io.deephaven.engine.util.PythonDeephavenSession;
+import io.deephaven.integrations.python.PythonDeephavenSession;
 import io.deephaven.engine.util.PythonScopeJpyImpl;
 import io.deephaven.engine.table.ColumnSource;
 import io.deephaven.jpy.PythonTest;
