+++ python-engine-test/src/test/java/io/deephaven/engine/util/WorkerPythonEnvironment.java
@@ -7,6 +7,8 @@ import io.deephaven.UncheckedDeephavenException;
 import io.deephaven.base.FileUtils;
 import io.deephaven.base.verify.Assert;
 import io.deephaven.configuration.Configuration;
+import io.deephaven.integrations.python.PythonDeephavenSession;
+import io.deephaven.integrations.python.PythonLogAdapter;
 import io.deephaven.internal.log.LoggerFactory;
 import io.deephaven.io.logger.Logger;
 import io.deephaven.engine.table.lang.QueryScope;
