+++ server/netty/build.gradle
@@ -5,6 +5,7 @@ plugins {
 
 dependencies {
     api project(':server')
+    compile project(':Integrations')
 
     Classpaths.inheritDagger(project)
     Classpaths.inheritDagger(project, /* test */ true)
