+++ server/src/main/java/io/deephaven/server/appmode/ApplicationFactory.java
@@ -4,17 +4,11 @@
 package io.deephaven.server.appmode;
 
 import com.google.rpc.Code;
-import io.deephaven.appmode.ApplicationConfig;
-import io.deephaven.appmode.ApplicationContext;
-import io.deephaven.appmode.ApplicationState;
-import io.deephaven.appmode.DynamicApplication;
-import io.deephaven.appmode.QSTApplication;
-import io.deephaven.appmode.ScriptApplication;
-import io.deephaven.appmode.StaticClassApplication;
+import io.deephaven.appmode.*;
 import io.deephaven.engine.util.GroovyDeephavenSession;
-import io.deephaven.engine.util.PythonDeephavenSession;
 import io.deephaven.engine.util.ScriptSession;
 import io.deephaven.extensions.barrage.util.GrpcUtil;
+import io.deephaven.integrations.python.PythonDeephavenSession;
 
 import java.nio.file.Path;
 import java.util.List;
