+++ server/src/test/java/io/deephaven/server/appmode/ApplicationTest.java
@@ -8,7 +8,7 @@ import io.deephaven.appmode.Field;
 import io.deephaven.engine.table.Table;
 import io.deephaven.engine.util.AbstractScriptSession;
 import io.deephaven.engine.util.GroovyDeephavenSession;
-import io.deephaven.engine.util.PythonDeephavenSession;
+import io.deephaven.integrations.python.PythonDeephavenSession;
 import io.deephaven.engine.util.PythonEvaluatorJpy;
 import io.deephaven.plugin.type.ObjectTypeLookup.NoOp;
 import io.deephaven.test.junit4.EngineCleanup;
